LPS-96877 Fix unit test
